+++ src/libs/Navigation/AppNavigator/Navigators/ReportsSplitNavigator.tsx
@@ -3,7 +3,7 @@ import React, {useRef} from 'react';
 import FocusTrapForScreens from '@components/FocusTrap/FocusTrapForScreen';
 import useActiveWorkspace from '@hooks/useActiveWorkspace';
 import usePermissions from '@hooks/usePermissions';
-import createSplitStackNavigator from '@libs/Navigation/AppNavigator/createSplitStackNavigator';
+import createSplitNavigator from '@libs/Navigation/AppNavigator/createSplitNavigator';
 import FreezeWrapper from '@libs/Navigation/AppNavigator/FreezeWrapper';
 import useRootNavigatorOptions from '@libs/Navigation/AppNavigator/useRootNavigatorOptions';
 import getCurrentUrl from '@libs/Navigation/currentUrl';
@@ -17,7 +17,7 @@ import type ReactComponentModule from '@src/types/utils/ReactComponentModule';
 const loadReportScreen = () => require<ReactComponentModule>('../../../../pages/home/ReportScreen').default;
 const loadSidebarScreen = () => require<ReactComponentModule>('@pages/home/sidebar/SidebarScreen').default;
 
-const Stack = createSplitStackNavigator<ReportsSplitNavigatorParamList>();
+const Split = createSplitNavigator<ReportsSplitNavigatorParamList>();
 
 function ReportsSplitNavigator() {
     const {canUseDefaultRooms} = usePermissions();
@@ -47,23 +47,23 @@ function ReportsSplitNavigator() {
     return (
         <FreezeWrapper>
             <FocusTrapForScreens>
-                <Stack.Navigator
+                <Split.Navigator
                     sidebarScreen={SCREENS.HOME}
                     defaultCentralScreen={SCREENS.REPORT}
                     parentRoute={route}
                     screenOptions={rootNavigatorOptions.centralPaneNavigator}
                 >
-                    <Stack.Screen
+                    <Split.Screen
                         name={SCREENS.HOME}
                         getComponent={loadSidebarScreen}
                         options={rootNavigatorOptions.homeScreen}
                     />
-                    <Stack.Screen
+                    <Split.Screen
                         name={SCREENS.REPORT}
                         initialParams={{reportID: initialReportID, openOnAdminRoom: shouldOpenOnAdminRoom() ? true : undefined}}
                         getComponent={loadReportScreen}
                     />
-                </Stack.Navigator>
+                </Split.Navigator>
             </FocusTrapForScreens>
         </FreezeWrapper>
     );
