+++ src/libs/Navigation/AppNavigator/createSplitNavigator/getInitialSplitNavigatorState.ts
