+++ src/libs/Navigation/AppNavigator/createSplitNavigator/useHandleScreenResize/index.native.ts
