+++ src/libs/Navigation/AppNavigator/createSplitNavigator/useHandleScreenResize/index.ts
