+++ src/libs/Navigation/NavigationRoot.tsx
@@ -22,7 +22,7 @@ import ONYXKEYS from '@src/ONYXKEYS';
 import type {Route} from '@src/ROUTES';
 import ROUTES from '@src/ROUTES';
 import AppNavigator from './AppNavigator';
-import {cleanPreservedSplitNavigatorStates} from './AppNavigator/createSplitStackNavigator/usePreserveSplitNavigatorState';
+import {cleanPreservedSplitNavigatorStates} from './AppNavigator/createSplitNavigator/usePreserveSplitNavigatorState';
 import {customGetPathFromState, getAdaptedStateFromPath, setupCustomAndroidBackHandler} from './helpers';
 import linkingConfig from './linkingConfig';
 import Navigation, {navigationRef} from './Navigation';
